Project/Backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
scikit-learn
pandas
joblib
flask
matplotlib
